updated the example
